requirements.txt
flask
flask-sock
openai
python-dotenv
requests
dashscope
